Use size_t as iterators for DAC loops, otherwise we create endless loops when scanning all the way up till 0xFF. Bump version to 1.26.

+++ software/dtb_expert/trigger_loops.cpp
@@ -176,7 +176,7 @@ void CTestboard::LoopMultiRocAllPixelsDacScan(vector<uint8_t> &roc_i2c, uint16_t
       }
 
       // Loop over the DAC range specified:
-      for (uint8_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
+      for (size_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
 
 	// Update the DAC setting on all configured ROCS:
 	for(size_t roc = 0; roc < roc_i2c.size(); roc++) {
@@ -221,7 +221,7 @@ void CTestboard::LoopMultiRocOnePixelDacScan(vector<uint8_t> &roc_i2c, uint8_t c
   }
 
   // Loop over the DAC range specified:
-  for (uint8_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
+  for (size_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
 
     // Update the DAC setting on all configured ROCS:
     for(size_t roc = 0; roc < roc_i2c.size(); roc++) {
@@ -266,7 +266,7 @@ void CTestboard::LoopSingleRocAllPixelsDacScan(uint8_t roc_i2c, uint16_t nTrigge
       roc_Pix_Cal(col, GetXtalkRow(row,(flags&FLAG_XTALK)), (flags&FLAG_CALS));
 
       // Loop over the DAC range specified:
-      for (uint8_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
+      for (size_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
 
 	// Update the DAC setting on the ROC:
 	roc_SetDAC(dac1register, dac1);
@@ -299,7 +299,7 @@ void CTestboard::LoopSingleRocOnePixelDacScan(uint8_t roc_i2c, uint8_t column, u
   roc_Pix_Cal(column, GetXtalkRow(row,(flags&FLAG_XTALK)), (flags&FLAG_CALS));
 
   // Loop over the DAC range specified:
-  for (uint8_t dac = dac1low; dac <= dac1high; dac++) {
+  for (size_t dac = dac1low; dac <= dac1high; dac++) {
 
     // Update the DAC setting on the ROC:
     roc_SetDAC(dac1register, dac);
@@ -342,7 +342,7 @@ void CTestboard::LoopMultiRocAllPixelsDacDacScan(vector<uint8_t> &roc_i2c, uint1
       }
 
       // Loop over the DAC1 range specified:
-      for (uint8_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
+      for (size_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
 
 	// Update the DAC1 setting on all configured ROCS:
 	for(size_t roc = 0; roc < roc_i2c.size(); roc++) {
@@ -351,7 +351,7 @@ void CTestboard::LoopMultiRocAllPixelsDacDacScan(vector<uint8_t> &roc_i2c, uint1
 	}
 
 	// Loop over the DAC2 range specified:
-	for (uint8_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
+	for (size_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
 
 	  // Update the DAC2 setting on all configured ROCS:
 	  for(size_t roc = 0; roc < roc_i2c.size(); roc++) {
@@ -397,7 +397,7 @@ void CTestboard::LoopMultiRocOnePixelDacDacScan(vector<uint8_t> &roc_i2c, uint8_
   }
 
   // Loop over the DAC1 range specified:
-  for (uint8_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
+  for (size_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
     
     // Update the DAC1 setting on all configured ROCS:
     for(size_t roc = 0; roc < roc_i2c.size(); roc++) {
@@ -406,7 +406,7 @@ void CTestboard::LoopMultiRocOnePixelDacDacScan(vector<uint8_t> &roc_i2c, uint8_
     }
 
     // Loop over the DAC2 range specified:
-    for (uint8_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
+    for (size_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
 
       // Update the DAC2 setting on all configured ROCS:
       for(size_t roc = 0; roc < roc_i2c.size(); roc++) {
@@ -452,13 +452,13 @@ void CTestboard::LoopSingleRocAllPixelsDacDacScan(uint8_t roc_i2c, uint16_t nTri
       roc_Pix_Cal(col, GetXtalkRow(row,(flags&FLAG_XTALK)), (flags&FLAG_CALS));
 
       // Loop over the DAC1 range specified:
-      for (uint8_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
+      for (size_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
     
 	// Update the DAC1 setting on the configured ROC:
 	roc_SetDAC(dac1register, dac1);
 
 	// Loop over the DAC2 range specified:
-	for (uint8_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
+	for (size_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
 
 	  // Update the DAC2 setting on the configured ROC:
 	  roc_SetDAC(dac2register, dac2);
@@ -492,13 +492,13 @@ void CTestboard::LoopSingleRocOnePixelDacDacScan(uint8_t roc_i2c, uint8_t column
   roc_Pix_Cal(column, GetXtalkRow(row,(flags&FLAG_XTALK)), (flags&FLAG_CALS));
 
   // Loop over the DAC range specified:
-  for (uint8_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
+  for (size_t dac1 = dac1low; dac1 <= dac1high; dac1++) {
     
     // Update the DAC1 setting on the configured ROC:
     roc_SetDAC(dac1register, dac1);
 
     // Loop over the DAC2 range specified:
-    for (uint8_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
+    for (size_t dac2 = dac2low; dac2 <= dac2high; dac2++) {
 
       // Update the DAC2 setting on the configured ROC:
       roc_SetDAC(dac2register, dac2);
